requirements.txt
astroid==2.5.6
autopep8==1.5.7
colorama==0.4.4
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pycodestyle==2.7.0
pylint==2.8.2
python-decouple==3.4
selenium==3.141.0
toml==0.10.2
typed-ast==1.4.3
urllib3==1.26.4
wrapt==1.12.1
